requirements.txt
Django==5.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
django-environ==0.12.0
bcrypt==4.2.1
sqlparse==0.5.2
colorama==0.4.6
cryptography==44.0.1
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-extensions==3.2.3
Faker==36.1.1
grpcio-status==1.70.0
gunicorn==23.0.0
pillow==11.1.0
pipdeptree==2.25.0
psycopg==3.2.4
redis==5.2.1
setuptools==65.5.0
watchdog==6.0.0
psycopg2-binary==2.9.10
flower
aamarpay